src/com/example/beer/api/ApiManager.java
package com.example.beer.api;

import org.apache.http.HttpResponse;
import org.apache.http.client.methods.HttpRequestBase;
import org.apache.http.impl.client.DefaultHttpClient;

import java.io.BufferedReader;
import java.io.IOException;
import java.io.InputStream;
import java.io.InputStreamReader;


public class ApiManager {

    public String executeRequest(final HttpRequestBase request) {
        try {
            DefaultHttpClient client = new DefaultHttpClient();
            HttpResponse response = client.execute(request);
            return readResponse(response);
        } catch (Exception ex) {
            return null;
        }
    }

    protected static String readResponse(HttpResponse response) throws IOException {
        StringBuilder builder = new StringBuilder();
        InputStream inputStream = response.getEntity().getContent();
        BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream));
        String line;
        while ((line = reader.readLine()) != null) {
            builder.append(line);
        }
        return builder.toString();
    }
}
